reorganize the distance function
