+++ src/components/AboutGame.jsx
@@ -1,0 +1,5 @@
+const AboutGame = () => {
+    return <h1>About the Games Afoot</h1>;
+};
+
+export default AboutGame;
