--- src/components/AboutUs.jsx
@@ -1,5 +1,0 @@
-const AboutUs = () => {
-    return <h1>About the developers</h1>;
-};
-
-export default AboutUs;
